==> real_estate/v9/00_get_postal_codes.ipynb <==
import pandas as pd

df = pd.read_table("~/real_estate/data/raw/cbp/zbp10totals.txt",sep=",")

county_df = df[["zip","cty_name","stabbr"]]
county_df = county_df.rename(columns={"zip":"postal_code",
                                      "cty_name":"county",
                                      "stabbr":"state"})
county_df["county_state"] = [str(i)+" "+str(j)for i,j in zip(df["cty_name"].values, df["stabbr"].values)]

county_df.to_csv("~/real_estate/data/v12/zipcode.csv",index=False)

county_df
