==> real_estate/v9/01_merge_dfs.ipynb <==
import pandas as pd

# Load data
zillow_df = pd.read_csv("~/real_estate/data/v4/unsupervised/zillow_zipcode_data.csv")
cbp_df = pd.read_csv("~/real_estate/data/v9/cbp_zipcode_df.csv")
irs_df = pd.read_csv("~/real_estate/data/v4/unsupervised/irs_zipcode_data.csv")
fed_df = pd.read_csv("~/real_estate/data/v4/unsupervised/fed_dataframe.csv")
postal_code_df = pd.read_csv("~/real_estate/data/v9/zipcode.csv")
postal_code_list = set(postal_code_df["postal_code"].values)

# Merge
df = zillow_df.loc[zillow_df["postal_code"].isin(postal_code_list)]
df = pd.merge(df, cbp_df, on=["postal_code","year"], how="left")
df = df.dropna(subset=["zhvi_condo"])
df = pd.merge(df, irs_df, on=["postal_code","year"], how="left")
df = pd.merge(df, fed_df, on=["year","month"], how="left")
#df = df.drop(["state"], axis=1)
#df = pd.merge(df, state_df[["postal_code","state"]], on="postal_code", how="left")

# Filter
df = df.dropna(subset=["state"])
df = df.loc[df["year"] >= 2010]

# Remove columns with tons of nas
df = df.drop(["zhvi_1bed","zhvi_2bed","zhvi_3bed"], axis=1)

# Sort
df = df.sort_values(by=["postal_code","year","month"])

print(df.shape)
df.head()

# Export
df.to_csv("~/real_estate/data/v9/unsupervised_df.csv",index=False)
